+++ src/main/scala/org/opennetworkinsight/DNSLDA.scala
@@ -6,23 +6,23 @@ import org.apache.spark.sql.SQLContext
 import org.apache.spark.SparkContext
 import org.slf4j.Logger
 
-object FlowLDA {
+object DNSLDA {
 
   def run(config: Config, sparkContext: SparkContext, sqlContext: SQLContext, logger: Logger) = {
 
-    logger.info("Flow LDA starts")
+    logger.info("DNS LDA starts")
 
-    val docWordCount = FlowPreLDA.flowPreLDA(config.inputPath, config.scoresFile, config.duplicationFactor, sparkContext,
+    val docWordCount = DNSPreLDA.dnsPreLDA(config.inputPath, config.scoresFile, config.duplicationFactor, sparkContext,
       sqlContext, logger)
-    // TODO persist docWordCount in MEMORY and DISK
+
     val ldaResult = LDAWrapper.runLDA(docWordCount, config.modelFile, config.topicDocumentFile, config.topicWordFile,
       config.mpiPreparationCmd, config.mpiCmd, config.mpiProcessCount, config.mpiTopicCount, config.localPath,
       config.localUser, config.ldaPath, config.dataSource, config.nodes)
 
-    FlowPostLDA.flowPostLDA(config.inputPath, config.hdfsScoredConnect, config.threshold, ldaResult("document_results"),
+    DNSPostLDA.dnsPostLDA(config.inputPath, config.hdfsScoredConnect, config.threshold, ldaResult("document_results"),
       ldaResult("word_results"), sparkContext, sqlContext, logger)
 
-    logger.info("Flow LDA completed")
+    logger.info("DNS LDA completed")
   }
 
 }
